feature: add password verifiers
In this commit we added validators to verify password integrities.

The validators are following these business rules https://github.com/itidigital/backend-challenge
which it's our acceptance criteria.

Also in this commit we added a principle to focus on domain layer
which is responsible to take all business logic and needs. This is giving to us
an ability to be indepentend of any layers in this project.